basic task app running
